slice client.tsx file
