+++ apps/web/src/app/admin/user/section/AdminManageConsole.tsx
@@ -1,0 +1,15 @@
+import { UserListByRole } from "./UserListByRole";
+import { PromoteTo } from "./PromoteTo";
+import { DemoteFrom } from "./DemoteFrom";
+
+export function AdminManageConsole (){
+    return (
+    <section>
+      <h2>ADMIV 관리</h2>
+        <UserListByRole Role="ADMIN"/>
+        <PromoteTo promoteRole="ADMIN"/>
+        <DemoteFrom demoteRole="ADMIN"/>
+    </section>
+  );
+}
+
